recursive getRecipeRequirements(), added sumAssemblers and sumRawItems
sumAssemblers() and sumRawItems() both will create a sub 
object as required, and if not, simply add its argument to the 
existing structure.

Has conditional rounding up, only if at the top level of the 
recursion. This is because the top level, or output Assemblers, 
will limit the rates of everything feeding them, so lower recursion 
depths can effectively have fractional assemblers worth of 
ingredient requirements.