Building raw optional UI
Added item.rawOptional array, which contains the 
specifc raw items which the UI will provide as 
optional.  If an item is in items.rawOptional, but is 
not in items.raw, then the UI will present that 
option as de-selected by default.

Need to add functions which add items to and from 
the raw array when they are checked or unchecked.
Also, all checkboxes should be given a common 
class to make parsing them easier.